+++ nuxeo-drive-client/nxdrive/osi/windows/windows.py
@@ -17,16 +17,19 @@ log = getLogger(__name__)
 class WindowsIntegration(AbstractOSIntegration):
 
     RUN_KEY = 'Software\\Microsoft\\Windows\\CurrentVersion\\Run'
+    class_key_base = 'Software\\Classes\\{}\\shell\\{}'
     __zoom_factor = None
 
     def __init__(self, manager):
         super(WindowsIntegration, self).__init__(manager)
 
-    def get_menu_parent_key(self):
-        return 'Software\\Classes\\*\\shell\\' + self._manager.app_name
+    @property
+    def file_key(self):
+        return self.class_key_base.format('*', self._manager.app_name)
 
-    def get_menu_key(self):
-        return self.get_menu_parent_key() + '\\command'
+    @property
+    def folder_key(self):
+        return self.class_key_base.format('directory', self._manager.app_name)
 
     @staticmethod
     def _delete_reg_value(reg, path, value):
@@ -40,16 +43,11 @@ class WindowsIntegration(AbstractOSIntegration):
     @staticmethod
     def _update_reg_key(reg, path, attributes=()):
         """ Helper function to create / set a key with attribute values. """
-
         key = _winreg.CreateKey(reg, path)
         _winreg.CloseKey(key)
         with _winreg.OpenKey(reg, path, 0, _winreg.KEY_WRITE) as key:
-            for attribute, type_, value in attributes:
-                # Handle None case for app name in
-                # contextual_menu.register_contextual_menu_win32
-                if attribute == 'None':
-                    attribute = None
-                _winreg.SetValueEx(key, attribute, 0, type_, value)
+            for name, type_, data in attributes:
+                _winreg.SetValueEx(key, name, 0, type_, data)
 
     @property
     def zoom_factor(self):
@@ -80,61 +78,54 @@ class WindowsIntegration(AbstractOSIntegration):
             return
 
         log.debug(
-            'Registering %r application %r to registry key %r',
-            app_name, exe_path, reg_key)
+            'Registering the executable %r to the registry key %r',
+            exe_path, reg_key)
         reg = _winreg.ConnectRegistry(None, _winreg.HKEY_CURRENT_USER)
         self._update_reg_key(
-            reg, reg_key,
-            [(app_name, _winreg.REG_SZ, exe_path)],
-        )
+            reg, reg_key, [(app_name, _winreg.REG_SZ, exe_path)])
 
     def unregister_startup(self):
         self._delete_reg_value(
             _winreg.HKEY_CURRENT_USER, self.RUN_KEY, self._manager.app_name)
 
     def register_protocol_handlers(self):
-        """Register ndrive as a protocol handler in the Registry"""
+        """ Register ndrive as a protocol handler in the Registry """
         exe_path = self._manager.find_exe_path()
-        app_name = self._manager.app_name
         if exe_path is None:
             return
+
+        app_name = self._manager.app_name
+        protocol_path = 'Software\\{}\\Protocols\\nxdrive'.format(app_name)
+        class_path = 'Software\\Classes\\nxdrive'
+        command = '"{}" "%1"'.format(exe_path)
 
         log.debug('Registering "nxdrive" protocol handler to: %r', exe_path)
         reg = _winreg.ConnectRegistry(None, _winreg.HKEY_CURRENT_USER)
 
         # Register Nuxeo Drive as a software as a protocol command provider
-        command = '"' + exe_path + '" "%1"'
         self._update_reg_key(
-            reg, 'Software\\' + app_name,
-            [('', _winreg.REG_SZ, app_name)],
-        )
+            reg, 'Software\\' + app_name, [('', _winreg.REG_SZ, app_name)])
+
         # TODO: add an icon for Nuxeo Drive too
         self._update_reg_key(
-            reg, 'Software\\' + app_name + '\\Protocols\\nxdrive',
-            [('URL Protocol', _winreg.REG_SZ, '')],
-        )
+            reg, protocol_path, [('URL Protocol', _winreg.REG_SZ, '')])
+
         # TODO: add an icon for the nxdrive protocol too
         self._update_reg_key(
-            reg,
-            'Software\\' + app_name + '\\Protocols\\nxdrive\\shell\\open\\command',
-            [('', _winreg.REG_SZ, command)],
-        )
+            reg, '{}\\shell\\open\\command'.format(protocol_path),
+            [('', _winreg.REG_SZ, command)])
+
         # Create the nxdrive protocol key
-        nxdrive_class_path = 'Software\\Classes\\nxdrive'
         self._update_reg_key(
-            reg, nxdrive_class_path,
-            [
+            reg, class_path, [
                 ('EditFlags', _winreg.REG_DWORD, 2),
                 ('', _winreg.REG_SZ, 'URL:nxdrive Protocol'),
-                ('URL Protocol', _winreg.REG_SZ, ''),
-            ],
-        )
+                ('URL Protocol', _winreg.REG_SZ, '')])
+
         # Create the nxdrive command key
-        command_path = nxdrive_class_path + '\\shell\\open\\command'
         self._update_reg_key(
-            reg, command_path,
-            [('', _winreg.REG_SZ, command)],
-        )
+            reg, '{}\\shell\\open\\command'.format(class_path),
+            [('', _winreg.REG_SZ, command)])
 
     def _recursive_delete(self, key0, key1, key2=''):
         """ Delete a key and its subkeys. """
@@ -144,9 +135,9 @@ class WindowsIntegration(AbstractOSIntegration):
             info = _winreg.QueryInfoKey(key)
             for x in range(info[0]):
                 """
-                Deleting the subkey will change the SubKey count used by EnumKey.
-                We must always pass 0 to EnumKey so we always get back the new
-                first SubKey.
+                Deleting the subkey will change the SubKey count
+                used by EnumKey. We must always pass 0 to EnumKey
+                so we always get back the new first SubKey.
                 """
                 subkey = _winreg.EnumKey(key, 0)
                 try:
@@ -162,29 +153,36 @@ class WindowsIntegration(AbstractOSIntegration):
     def unregister_protocol_handlers(self):
         app_name = self._manager.app_name
         reg = _winreg.HKEY_CURRENT_USER
-        self._recursive_delete(reg, 'Software\\' + app_name + '\\Protocols\\nxdrive')
+        self._recursive_delete(
+            reg, 'Software\\{}\\Protocols\\nxdrive'.format(app_name))
         self._recursive_delete(reg, 'Software\\Classes\\nxdrive')
 
     def register_contextual_menu(self):
-        # TODO: better understand why / how this works.
-        # See https://jira.nuxeo.com/browse/NXDRIVE-120
-        app_name = 'None'
-        args = ' metadata --file "%1"'
-        exe_path = self._manager.find_exe_path() + args
-        if exe_path is None:
-            return
-        icon_path = self._manager.find_exe_path() + ',0'
-        log.debug('Registering %r application %r to registry key %r',
-                  app_name, exe_path, self.get_menu_key())
+        """
+        Modify the registry to add an entry in the contextual menu.
+
+        When adding an entry in a registry key, passing `None` as
+        the name will set the `default` value of the key.
+        That way, the Windows Explorer sees the right label and
+        command to add as an entry in the contextual menu.
+        """
+        exe_path = self._manager.find_exe_path()
+        menu_cmd = '{} metadata --file "%1"'.format(exe_path)
+        icon_path = '{},0'.format(exe_path)
+        label = 'Access online'
+
         reg = _winreg.ConnectRegistry(None, _winreg.HKEY_CURRENT_USER)
-        self._update_reg_key(
-            reg, self.get_menu_key(),
-            [(app_name, _winreg.REG_SZ, exe_path)],
-        )
-        self._update_reg_key(
-            reg, self.get_menu_parent_key(),
-            [("Icon", _winreg.REG_SZ, icon_path)],
-        )
+
+        # Add a key for both file and folder context menus
+        for key in (self.file_key, self.folder_key):
+            subkey = '{}\\command'.format(key)
+            log.debug('Registering the command `%r` to the registry key %r',
+                      menu_cmd, subkey)
+            self._update_reg_key(reg, subkey,
+                                 [(None, _winreg.REG_SZ, menu_cmd)])
+            self._update_reg_key(reg, key,
+                                 [(None, _winreg.REG_SZ, label),
+                                  ('Icon', _winreg.REG_SZ, icon_path)])
 
     @staticmethod
     def is_same_partition(folder1, folder2):
@@ -204,8 +202,9 @@ class WindowsIntegration(AbstractOSIntegration):
     def get_system_configuration(self):
         result = dict()
         reg = _winreg.HKEY_CURRENT_USER
+        reg_key = 'Software\\Nuxeo\\Drive'
         try:
-            with _winreg.OpenKey(reg, 'Software\\Nuxeo\\Drive', 0, _winreg.KEY_READ) as key:
+            with _winreg.OpenKey(reg, reg_key, 0, _winreg.KEY_READ) as key:
                 for i in range(_winreg.QueryInfoKey(key)[1]):
                     k, v, _ = _winreg.EnumValue(key, i)
                     result[k.replace('-', '_').lower()] = v
@@ -215,9 +214,11 @@ class WindowsIntegration(AbstractOSIntegration):
 
     def unregister_contextual_menu(self):
         reg = _winreg.HKEY_CURRENT_USER
-        if self._delete_reg_value(reg, self.get_menu_key(), ''):
-            _winreg.DeleteKey(reg, self.get_menu_key())
-            _winreg.DeleteKey(reg, self.get_menu_parent_key())
+        for key in (self.file_key, self.folder_key):
+            subkey = '{}\\command'.format(key)
+            if self._delete_reg_value(reg, subkey, ''):
+                _winreg.DeleteKey(reg, subkey)
+                _winreg.DeleteKey(reg, key)
 
     def register_folder_link(self, folder_path, name=None):
         favorite = self._get_folder_link(name)